Improve UX for Spotify quit command in osx plugin (#5726)
* Improve UX for Spotify commands in osx plugin

Spotify command UX is a bit weak for people who live and work almost entirely in shells.
- `spotify quit` should not open Spotify if it is not already running. Should confirm that Spotify is indeed not running.
- `spotify quit` should not blow away the user's shell once Spotify is quit. This can be a disruption to work flow.

This PR looks to add a few little checks which will help improve this experience.
This PR also adds a space to line 477 between `break` and `;;` for consistency. Doesn't seem like a big enough change to put in its own PR.

* Rearranging output as per peer feedback

* osx plugin's spotify: change quitting w/ closing

Closing is more idiomatic English.